kook/requirements.txt
dropbox==11.27.0
python-dotenv==0.15.0
python-slugify==6.0.1
